requirements.txt
streamlit
sentence-transformers
google-generativeai
python-dotenv
numpy
pandas
